Refactor RGB_LED and console blocks to individual modules.
