+++ src/modules/editor.js
@@ -31,6 +31,8 @@ import {tooltipText} from './blockly/language/en/messages';
 import /* webpackPrefetch: true */ './blockly/generators/propc';
 import './blockly/generators/propc/base';
 import './blockly/generators/propc/comms/i2c_protocol';
+import './blockly/generators/propc/console';
+import './blockly/generators/propc/rgb_led';
 import './blockly/generators/propc/communicate';
 import './blockly/generators/propc/control';
 import './blockly/generators/propc/cogs';
